Remove dependency on noflet
I think `cl-flet` is sufficient for `with-ansi` because we only want to
bind the color names to the functions lexically, not dynamically.

In other words we don't want/need it to work for things like this:
```
(defun foo ()
  (black "foo"))

(with-ansi (foo))
```

In fact it seems like using dynamic flet could even be dangerous here.
For example imagine a user had their own function called `black` that they
intended to call from `foo` in the code above.